+++ packages/core/src/initialVisit.ts
@@ -36,7 +36,7 @@ export class InitialVisit {
     history
       .decrypt()
       .then((data) => {
-        currentPage.set(data, { preserveScroll: true, preserveState: true, replace: true }).then(() => {
+        currentPage.set(data, { forgetScroll: false, forgetState: false, replace: true }).then(() => {
           Scroll.restore(currentPage.get())
           fireNavigateEvent(currentPage.get())
         })
@@ -76,11 +76,11 @@ export class InitialVisit {
 
         currentPage
           .set(currentPage.get(), {
-            preserveScroll: locationVisit.preserveScroll,
-            preserveState: true,
+            forgetScroll: locationVisit.forgetScroll,
+            forgetState: false,
           })
           .then(() => {
-            if (locationVisit.preserveScroll) {
+            if (locationVisit.forgetScroll) {
               Scroll.restore(currentPage.get())
             }
 
@@ -99,7 +99,7 @@ export class InitialVisit {
       currentPage.setUrlHash(window.location.hash)
     }
 
-    currentPage.set(currentPage.get(), { preserveState: true, replace: true }).then(() => {
+    currentPage.set(currentPage.get(), { forgetState: false, replace: true }).then(() => {
       fireNavigateEvent(currentPage.get())
     })
   }