+++ packages/svelte/test-app/Pages/Links/PreserveScroll.svelte
@@ -26,7 +26,7 @@
 
   <a
     href="/links/preserve-scroll-page-two"
-    use:inertia={{ preserveScroll: true, data: { foo: 'baz' } }}
+    use:inertia={{ forgetScroll: false, data: { foo: 'baz' } }}
     class="preserve"
     data-testid="preserve"
   >
@@ -38,7 +38,7 @@
 
   <a
     href="/links/preserve-scroll-page-two"
-    use:inertia={{ preserveScroll: preserveCallback, data: { foo: 'baz' } }}
+    use:inertia={{ forgetScroll: preserveCallback, data: { foo: 'baz' } }}
     class="preserve-callback"
     data-testid="preserve-callback"
   >
@@ -46,7 +46,7 @@
   </a>
   <a
     href="/links/preserve-scroll-page-two"
-    use:inertia={{ preserveScroll: preserveCallbackFalse, data: { foo: 'foo' } }}
+    use:inertia={{ forgetScroll: preserveCallbackFalse, data: { foo: 'foo' } }}
     class="reset-callback"
     data-testid="reset-callback"
   >