+++ CHANGELOG.md
@@ -33,7 +33,7 @@ For changes prior to v1.0.0, see the [legacy releases](https://legacy.inertiajs.
 ### Changed
 
 - Skip intercepting non-left button clicks on links ([#1908](https://github.com/inertiajs/inertia/pull/1908), [#1910](https://github.com/inertiajs/inertia/pull/1910))
-- Changed `preserveScroll` to be `true` on initial page visit ([#1360](https://github.com/inertiajs/inertia/pull/1360))
+- Changed `forgetScroll` to be `true` on initial page visit ([#1360](https://github.com/inertiajs/inertia/pull/1360))
 - Return early when using `router.on()` during SSR ([#1715](https://github.com/inertiajs/inertia/pull/1715))
 - Use updater function in `setData` in `useForm` hook in React adapter ([#1859](https://github.com/inertiajs/inertia/pull/1859))
 
@@ -50,12 +50,12 @@ For changes prior to v1.0.0, see the [legacy releases](https://legacy.inertiajs.
 - Fix type augmentation in Vue 3 adapter ([#1958](https://github.com/inertiajs/inertia/pull/1958))
 - Fix form helper `transform` return type in React adapter ([#1896](https://github.com/inertiajs/inertia/pull/1896))
 - Fix props reactivity in Svelte adapter ([#1969](https://github.com/inertiajs/inertia/pull/1969))
-- Fix `<Render />` component to respect `preserveState` option in Svelte adapter ([#1943](https://github.com/inertiajs/inertia/pull/1943))
+- Fix `<Render />` component to respect `forgetState` option in Svelte adapter ([#1943](https://github.com/inertiajs/inertia/pull/1943))
 - Fix 'received an unexpected slot "default"' warning in Svelte adapter ([#1941](https://github.com/inertiajs/inertia/pull/1941))
 
 ## [v1.2.0](https://github.com/inertiajs/inertia/compare/v1.1.0...v1.2.0)
 
-- Fix `preserveScroll` and `preserveState` types ([#1882](https://github.com/inertiajs/inertia/pull/1882))
+- Fix `forgetScroll` and `forgetState` types ([#1882](https://github.com/inertiajs/inertia/pull/1882))
 - Revert "merge props from partial reloads" ([#1895](https://github.com/inertiajs/inertia/pull/1895))
 
 ## [v1.1.0](https://github.com/inertiajs/inertia/compare/v1.0.16...v1.1.0)